Create a mft_analyzer.py, move parse_records and get_filepath to it, remove the mft_testing.py as the features are integrated.
