+++ analyzemft/mftsession.py
@@ -156,36 +156,7 @@ class MftSession:
                if self.options.bodyfile != None:
                     self.file_body.write(mft.mft_to_body(self.mft[i], self.options.bodyfull, self.options.bodystd))
 
-     def get_folder_path(self, seqnum):
-          if self.debug: print  ("Building Folder For Record Number (%d)" % seqnum)
 
-          if seqnum not in self.mft:
-               return 'Orphan'
-
-          # If we've already figured out the path name, just return it
-          if (self.mft[seqnum]['filename']) != '':
-               return self.mft[seqnum]['filename']
-
-          try:
-#                if (self.mft[seqnum]['fn',0]['par_ref'] == 0) or (self.mft[seqnum]['fn',0]['par_ref'] == 5):  # There should be no seq number 0, not sure why I had that check in place.
-               if (self.mft[seqnum]['fn',0]['par_ref'] == 5): # Seq number 5 is "/", root of the directory
-                    self.mft[seqnum]['filename'] = '/' + self.mft[seqnum]['fn',self.mft[seqnum]['fncnt']-1]['name']
-                    return self.mft[seqnum]['filename']
-          except:  # If there was an error getting the parent's sequence number, then there is no FN record
-               self.mft[seqnum]['filename'] = 'NoFNRecord'
-               return self.mft[seqnum]['filename']
-
-          # Self referential parent sequence number. The filename becomes a NoFNRecord note
-          if (self.mft[seqnum]['fn',0]['par_ref']) == seqnum:
-               if self.debug: print  ("Error, self-referential, while trying to determine path for seqnum %s" % seqnum)
-               self.mft[seqnum]['filename'] = 'ORPHAN/' + self.mft[seqnum]['fn',self.mft[seqnum]['fncnt']-1]['name']
-               return self.mft[seqnum]['filename']
-
-          # We're not at the top of the tree and we've not hit an error
-          parentpath = self.get_folder_path((self.mft[seqnum]['fn',0]['par_ref']))
-          self.mft[seqnum]['filename'] =  parentpath + '/' + self.mft[seqnum]['fn',self.mft[seqnum]['fncnt']-1]['name']
-
-          return self.mft[seqnum]['filename']
 
 
      def gen_filepaths(self):
